feat: Enable pool kubeconfig pattern
Allow the concourse-helm-resource to be used in put-only mode with a
kubeconfig file.

This builds off of pr46, but extends it so that there is no need to
configure the resource with the cluster_url (or anything else specified
in the kubeconfig).

Additionally, add a number of additional related controls:

- Get the k8s namespace from the kubeconfig file
- Ability to configure the tiller namespace
- Ability to enable tracing for debugging the resource
- Expose the helm init --wait option
- Ability to specify the kubeconfig path on put.

Together, these features allow us to use the helm resource in conjunction
with the pool resource and the kubernetes resource to:

- Dynamically pick a k8s namespace from a pool, granting exclusive access
- Run a helm-based standalone integration test against a clean environment
  in a dedicated namespace with a dedicated tiller
  - The concourse-helm-resource cleans up anything that may have been in the
    namespace's tiller
  - The concourse-helm-resource installs helm into the namespace
  - The concourse-helm-resource waits for tiller to be ready
  - The concourse-helm-resource installs the chart into the namespace with
    the namespace-specific tiller
- Gather the results (with the kubernetes resource)
- Return the k8s namespace to the pool

+++ assets/common.sh
@@ -4,45 +4,48 @@ set -e
 setup_kubernetes() {
   payload=$1
   source=$2
-  # Setup kubectl
-  cluster_url=$(jq -r '.source.cluster_url // ""' < $payload)
-  if [ -z "$cluster_url" ]; then
-    echo "invalid payload (missing cluster_url)"
-    exit 1
-  fi
-  if [[ "$cluster_url" =~ https.* ]]; then
 
-    cluster_ca=$(jq -r '.source.cluster_ca // ""' < $payload)
-    admin_key=$(jq -r '.source.admin_key // ""' < $payload)
-    admin_cert=$(jq -r '.source.admin_cert // ""' < $payload)
-    token=$(jq -r '.source.token // ""' < $payload)
-    token_path=$(jq -r '.params.token_path // ""' < $payload)
-
+  kubeconfig_path=$(jq -r '.params.kubeconfig_path // ""' < $payload)
+  absolute_kubeconfig_path="${source}/${kubeconfig_path}"
+  if [ -f "$absolute_kubeconfig_path" ]; then
     mkdir -p /root/.kube
+    cp "$absolute_kubeconfig_path" "/root/.kube/config"
+  else
+    # Setup kubectl
+    cluster_url=$(jq -r '.source.cluster_url // ""' < $payload)
+    if [ -z "$cluster_url" ]; then
+      echo "invalid payload (missing cluster_url)"
+      exit 1
+    fi
+    if [[ "$cluster_url" =~ https.* ]]; then
 
-    ca_path="/root/.kube/ca.pem"
-    echo "$cluster_ca" | base64 -d > $ca_path
-    kubectl config set-cluster default --server=$cluster_url --certificate-authority=$ca_path
+      cluster_ca=$(jq -r '.source.cluster_ca // ""' < $payload)
+      admin_key=$(jq -r '.source.admin_key // ""' < $payload)
+      admin_cert=$(jq -r '.source.admin_cert // ""' < $payload)
+      token=$(jq -r '.source.token // ""' < $payload)
+      token_path=$(jq -r '.params.token_path // ""' < $payload)
 
-    if [ -f "$source/$token_path" ]; then
-      kubectl config set-credentials admin --token=$(cat $source/$token_path)
-    elif [ ! -z "$token" ]; then
-      kubectl config set-credentials admin --token=$token
+      if [ -f "$source/$token_path" ]; then
+        kubectl config set-credentials admin --token=$(cat $source/$token_path)
+      elif [ ! -z "$token" ]; then
+        kubectl config set-credentials admin --token=$token
+      else
+        key_path="/root/.kube/key.pem"
+        cert_path="/root/.kube/cert.pem"
+        echo "$admin_key" | base64 -d > $key_path
+        echo "$admin_cert" | base64 -d > $cert_path
+        kubectl config set-credentials admin --client-certificate=$cert_path --client-key=$key_path
+      fi
+
+      kubectl config set-context default --cluster=default --user=admin
     else
-      key_path="/root/.kube/key.pem"
-      cert_path="/root/.kube/cert.pem"
-      echo "$admin_key" | base64 -d > $key_path
-      echo "$admin_cert" | base64 -d > $cert_path
-      kubectl config set-credentials admin --client-certificate=$cert_path --client-key=$key_path
+      kubectl config set-cluster default --server=$cluster_url
+      kubectl config set-context default --cluster=default
     fi
 
-    kubectl config set-context default --cluster=default --user=admin
-  else
-    kubectl config set-cluster default --server=$cluster_url
-    kubectl config set-context default --cluster=default
+    kubectl config use-context default
   fi
 
-  kubectl config use-context default
   kubectl version
 }
 
@@ -81,7 +84,15 @@ setup_tls() {
 
 setup_helm() {
   init_server=$(jq -r '.source.helm_init_server // "false"' < $1)
-  tiller_namespace=$(jq -r '.source.tiller_namespace // "kube-system"' < $1)
+
+  kubeconfig_tiller_namespace=$(jq -r '.source.kubeconfig_tiller_namespace // "false"' <$1)
+  if [ "$kubeconfig_tiller_namespace" = "true" ]
+  then
+    tiller_namespace=$(kubectl config view --minify -ojson | jq -r .contexts[].context.namespace)
+  else
+    tiller_namespace=$(jq -r '.source.tiller_namespace // "kube-system"' < $1)
+  fi
+
   tillerless=$(jq -r '.source.tillerless // "false"' < $payload)
   tls_enabled=$(jq -r '.source.tls_enabled // "false"' < $payload)
   history_max=$(jq -r '.source.helm_history_max // "0"' < $1)
@@ -105,6 +116,13 @@ setup_helm() {
       exit 1
     fi
     tiller_service_account=$(jq -r '.source.tiller_service_account // "default"' < $1)
+
+    helm_init_wait=$(jq -r '.source.helm_init_wait // "false"' <$1)
+    helm_init_wait_arg=""
+    if [ "$helm_init_wait" = "true" ]; then
+      helm_init_wait_arg="--wait"
+    fi
+
     if [ "$tls_enabled" = true ]; then
       tiller_key=$(jq -r '.source.tiller_key // ""' < $payload)
       tiller_cert=$(jq -r '.source.tiller_cert // ""' < $payload)
@@ -121,19 +139,29 @@ setup_helm() {
       helm_ca_cert_path="/root/.helm/ca.pem"
       echo "$tiller_key" > $tiller_key_path
       echo "$tiller_cert" > $tiller_cert_path
-      $helm_bin init --tiller-tls --tiller-tls-cert $tiller_cert_path --tiller-tls-key $tiller_key_path --tiller-tls-verify --tls-ca-cert $tiller_key_path --tiller-namespace=$tiller_namespace --service-account=$tiller_service_account --history-max=$history_max $stable_repo --upgrade
+      $helm_bin init --tiller-tls --tiller-tls-cert $tiller_cert_path --tiller-tls-key $tiller_key_path --tiller-tls-verify --tls-ca-cert $tiller_key_path --tiller-namespace=$tiller_namespace --service-account=$tiller_service_account --history-max=$history_max $stable_repo --upgrade $helm_init_wait_arg
     else
-      $helm_bin init --tiller-namespace=$tiller_namespace --service-account=$tiller_service_account --history-max=$history_max $stable_repo --upgrade
+      $helm_bin init --tiller-namespace=$tiller_namespace --service-account=$tiller_service_account --history-max=$history_max $stable_repo --upgrade $helm_init_wait_arg
     fi
     wait_for_service_up tiller-deploy 10
   else
     export HELM_HOST=$(jq -r '.source.helm_host // ""' < $1)
     $helm_bin init -c --tiller-namespace $tiller_namespace $stable_repo > /dev/null
   fi
+
+  tls_enabled_arg=""
   if [ "$tls_enabled" = true ]; then
-    $helm_bin version --tls --tiller-namespace $tiller_namespace
-  else
-    $helm_bin version --tiller-namespace $tiller_namespace
+    tls_enabled_arg="--tls"
+  fi
+  $helm_bin version $tls_enabled_arg --tiller-namespace $tiller_namespace
+
+  helm_setup_purge_all=$(jq -r '.source.helm_setup_purge_all // "false"' <$1)
+  if [ "$helm_setup_purge_all" = "true" ]; then
+    local release
+    for release in $(helm ls -aq --tiller-namespace $tiller_namespace )
+    do
+      helm delete $tls_enabled_arg --purge "$release" --tiller-namespace $tiller_namespace
+    done
   fi
 }
 
@@ -154,7 +182,14 @@ wait_for_service_up() {
 setup_repos() {
   repos=$(jq -c '(try .source.repos[] catch [][])' < $1)
   plugins=$(jq -c '(try .source.plugins[] catch [][])' < $1)
-  tiller_namespace=$(jq -r '.source.tiller_namespace // "kube-system"' < $1)
+
+  kubeconfig_tiller_namespace=$(jq -r '.source.kubeconfig_tiller_namespace // "false"' <$1)
+  if [ "$kubeconfig_tiller_namespace" = "true" ]
+  then
+    tiller_namespace=$(kubectl config view --minify -ojson | jq -r .contexts[].context.namespace)
+  else
+    tiller_namespace=$(jq -r '.source.tiller_namespace // "kube-system"' < $1)
+  fi
 
   local IFS=$'\n'
 
@@ -185,6 +220,11 @@ setup_repos() {
 }
 
 setup_resource() {
+  tracing_enabled=$(jq -r '.source.tracing_enabled // "false"' < $1)
+  if [ "$tracing_enabled" = "true" ]; then
+    set -x
+  fi
+
   echo "Initializing kubectl..."
   setup_kubernetes $1 $2
   echo "Initializing helm..."
